% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b.R
\name{b}
\alias{b}
\title{Extract regression coefficients from a linear regression model.}
\usage{
b(object, index = NULL, data = NULL)
}
\arguments{
\item{object}{An R object that is a formula or contains the results of the `lm` function.}

\item{index}{Index indicating which coefficients to obtain.}

\item{data}{A data frame when the object is a formula.}
}
\description{
The regression coefficients demonstrates how a set of predictor
variables will affect the outcome of interest.
}
