% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{descriptive}
\alias{descriptive}
\title{Obtain Descriptive Statistics from a data frame.}
\usage{
descriptive(df)
}
\arguments{
\item{df}{An R data frame used for further analysis.}
}
\description{
Obtain Descriptive Statistics from a data frame.
}
