% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.R
\name{f}
\alias{f}
\title{Extract the F Statistic}
\usage{
f(object)
}
\arguments{
\item{object}{An R object that is a `formula` or contains the results of the `lm` function.}
}
\description{
Extract the F Statistic
}
