% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{geom_pie}
\alias{geom_pie}
\title{A function that creates a pie chart.}
\usage{
geom_pie(mapping = NULL, data = NULL, position = "stack", ..., width = 1)
}
\arguments{
\item{mapping}{Aesthetics created by aes().}

\item{data}{The data used to create the plot.}

\item{position}{Set to "stack" for the pie chart.}

\item{...}{A set of arguments passed down to geom_bar().}

\item{width}{Set to 1 for the pie chart.}
}
\value{
A ggplot from the displays the pie chart.
}
\description{
A function that creates a pie chart.
}
