% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_info.R
\name{linear_model_info}
\alias{linear_model_info}
\title{Extract model information for linear regression.}
\usage{
linear_model_info(model)
}
\arguments{
\item{model}{An R object that results from a linear regression model.}
}
\value{
A message model information.
}
\description{
Extract model information for linear regression.
}
