% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_info.R
\name{logistic_model_info}
\alias{logistic_model_info}
\title{Extract model information for logistic regression}
\usage{
logistic_model_info(model)
}
\arguments{
\item{model}{An R object that results from a logistic regression model (glm class).}
}
\value{
A message indicating which category is being modeled as success from the logistic regression model.
}
\description{
Extract model information for logistic regression
}
