% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{num_by_cat_stats}
\alias{num_by_cat_stats}
\title{Obtain Numerical Statistics for a Continuous Variable by a Categorical Variable}
\usage{
num_by_cat_stats(df, num, cat)
}
\arguments{
\item{df}{An R data frame used for further analysis.}

\item{num}{Name of the numerical variable found in the data frame}

\item{cat}{Name of the categorical variable found in the data frame}
}
\description{
Obtain Numerical Statistics for a Continuous Variable by a Categorical Variable
}
