% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive.R
\name{num_stats}
\alias{num_stats}
\title{Obtain Numerical Statistics for a Continuous Variable}
\usage{
num_stats(x, tbl = TRUE, digits = 3)
}
\arguments{
\item{x}{A numerical or integer vector.}

\item{tbl}{A logical indicating whether to return a tibble or not, defaults to TRUE.}

\item{digits}{number of decimal places to round}
}
\description{
Obtain Numerical Statistics for a Continuous Variable
}
