% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/props.R
\name{props}
\alias{props}
\title{props Computing proportions}
\usage{
props(x, y, yval, diff = FALSE)
}
\arguments{
\item{x}{grouping variable}

\item{y}{outcome of interest}

\item{yval}{category of interest from outcome of interest}

\item{diff}{Obtain the difference in proportions. Default is FALSE.}
}
\description{
Compute the proportions of observing a value from a 2 by 2 continguency table
}
