% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2.R
\name{r2}
\alias{r2}
\title{Obtain the R-Squared Value from a Linear Model}
\usage{
r2(object)
}
\arguments{
\item{object}{An R object that is a `formula` or contains the results of the `lm` function.}
}
\description{
Obtain the R-Squared Value from a Linear Model
}
