% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcistats-package.R
\docType{package}
\name{rcistats-package}
\alias{rcistats}
\alias{rcistats-package}
\title{rcistats: R Tools for CSU Channel Islands}
\description{
An R package containing functions for statistics courses at CSUCI.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://www.inqs.info/rcistats/}
}

}
\keyword{internal}
