% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample.R
\name{resample}
\alias{resample}
\title{Sample a data frame with replacement}
\usage{
resample(df)
}
\arguments{
\item{df}{Data frame to be sampled with replacement}
}
\description{
Sample a data frame with replacement
}
