% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b.R
\name{se_b}
\alias{se_b}
\title{Extract the standard errors of the regression coefficients from a linear regression model.}
\usage{
se_b(object, index = NULL, data = NULL)
}
\arguments{
\item{object}{An R object that is a formula or contains the results of the `lm` function.}

\item{index}{Index indicating which coefficients to obtain.}

\item{data}{A data frame when the object is a formula.}
}
\description{
The standard errors of the regression coefficients demonstrates the variability of the relationship
between the predictor variables and the outcome of interest.
}
