% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sse.R
\name{sse}
\alias{sse}
\title{Compute the sum of error squared for an R object}
\usage{
sse(object, data = NULL)
}
\arguments{
\item{object}{An R object that is a `formula` or contains the results of the `lm` function.}

\item{data}{A data frame when the object is a formula.}
}
\description{
Compute the sum of error squared for an R object
}
