% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unicorns.R
\name{unicorns}
\alias{unicorns}
\title{Creates a Unicorn Data Set from simulated models.}
\usage{
unicorns(n)
}
\arguments{
\item{n}{Number of unicorns used}
}
\description{
Creates a Unicorn Data Set from simulated models.
}
